requirements.txt
defusedxml==0.5.0
Django==1.10.5
django-allauth==0.30.0
django-guardian==1.4.6
django-summernote==0.8.6
django-taggit==0.22.0
oauthlib==2.0.1
olefile==0.44
Pillow==4.0.0
pkg-resources==0.0.0
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
